ai_service/requirements.txt
fastapi
"uvicorn[standard]"
torch
librosa
faiss-cpu
numpy

